requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.0
python-dotenv==1.0.1
Pillow==10.4.0
supabase==2.28.3
